requirements.txt
greenlet==3.0.0
playwright==1.39.0
pyee==11.0.1
python-dotenv==1.0.0
typing_extensions==4.8.0
